refactor: search.js 공통함수 분리 - 콜백함수로 분리 - index.js에서 search 함수 호출: filterFn 매개변수로 넘김 - search 함수 내부에서 filter 함수 실행 item과 관련된 keyword 매칭
